CLOUD-827 relocate gcp domain objects from 'service' package
