+++ core-model/src/main/java/com/sequenceiq/cloudbreak/domain/GcpDiskType.java
@@ -1,4 +1,4 @@
-package com.sequenceiq.cloudbreak.service.stack.connector.gcp.domain;
+package com.sequenceiq.cloudbreak.domain;
 
 public enum GcpDiskType {
