+++ core/src/main/java/com/sequenceiq/cloudbreak/service/price/GCPPriceGenerator.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 import com.sequenceiq.cloudbreak.domain.CloudPlatform;
 import com.sequenceiq.cloudbreak.domain.GcpTemplate;
 import com.sequenceiq.cloudbreak.domain.Template;
-import com.sequenceiq.cloudbreak.service.stack.connector.gcp.domain.GcpInstanceType;
+import com.sequenceiq.cloudbreak.domain.GcpInstanceType;
 
 @Component
 public class GCPPriceGenerator implements PriceGenerator {
