+++ core/src/main/java/com/sequenceiq/cloudbreak/service/stack/resource/gcp/builders/instance/GcpInstanceResourceBuilder.java
@@ -43,9 +43,9 @@ import com.sequenceiq.cloudbreak.service.stack.connector.gcp.GcpResourceCheckerS
 import com.sequenceiq.cloudbreak.service.stack.connector.gcp.GcpResourceReadyPollerObject;
 import com.sequenceiq.cloudbreak.service.stack.connector.gcp.GcpStackUtil;
 import com.sequenceiq.cloudbreak.service.stack.connector.gcp.GcpResourceException;
-import com.sequenceiq.cloudbreak.service.stack.connector.gcp.domain.GcpDiskMode;
-import com.sequenceiq.cloudbreak.service.stack.connector.gcp.domain.GcpDiskType;
-import com.sequenceiq.cloudbreak.service.stack.connector.gcp.domain.GcpZone;
+import com.sequenceiq.cloudbreak.domain.GcpDiskMode;
+import com.sequenceiq.cloudbreak.domain.GcpDiskType;
+import com.sequenceiq.cloudbreak.domain.GcpZone;
 import com.sequenceiq.cloudbreak.service.stack.resource.CreateResourceRequest;
 import com.sequenceiq.cloudbreak.service.stack.resource.gcp.GcpSimpleInstanceResourceBuilder;
 import com.sequenceiq.cloudbreak.service.stack.resource.gcp.model.GcpDeleteContextObject;
