+++ core/src/main/java/com/sequenceiq/cloudbreak/service/stack/resource/gcp/builders/network/GcpReservedIpResourceBuilder.java
@@ -25,7 +25,7 @@ import com.sequenceiq.cloudbreak.service.stack.connector.gcp.GcpRemoveReadyPolle
 import com.sequenceiq.cloudbreak.service.stack.connector.gcp.GcpResourceCheckerStatus;
 import com.sequenceiq.cloudbreak.service.stack.connector.gcp.GcpResourceReadyPollerObject;
 import com.sequenceiq.cloudbreak.service.stack.connector.gcp.GcpResourceException;
-import com.sequenceiq.cloudbreak.service.stack.connector.gcp.domain.GcpZone;
+import com.sequenceiq.cloudbreak.domain.GcpZone;
 import com.sequenceiq.cloudbreak.service.stack.resource.CreateResourceRequest;
 import com.sequenceiq.cloudbreak.service.stack.resource.gcp.GcpSimpleNetworkResourceBuilder;
 import com.sequenceiq.cloudbreak.service.stack.resource.gcp.model.GcpDeleteContextObject;
